requirements.txt
streamlit==1.37.1
python-docx==1.1.2
pydantic==2.8.2
rapidfuzz==3.9.7
pyyaml==6.0.2
reportlab==4.2.2
pytest==8.3.2
fastapi==0.112.2
uvicorn[standard]==0.30.5
python-multipart==0.0.9
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
